+++ CHANGELOG.md
@@ -7,6 +7,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Added
 - `fusion` method for tilings based on the `Fusion` class.
+- `row_and_column_separation` method for tilings based on the
+  `RowColSeperation` class.
+- Adding `Factor`, `FactorWithInterleaving` and
+  `FactorWithMonotoneInterleaving` algorithm. They are available by setting the
+  right `interleaving` flag in the `find_factors` method of tilings
+
+### Changed
+- Update permuta to 1.2.1
+- Update comb_spec_searcher to 0.2.2
 
 ## [1.0.1] - 2019-08-26
 ### Changed